Move remote-config typings to another file
